centralize and share types moar betterer
