+++ assets/src/domain/eventEditor/interfaces/types.ts
@@ -1,5 +1,5 @@
 import { RelationalData } from '../../../application/services/apollo/relations/types';
-import { EventEspressoDomData } from '../../../application/services/config/types';
+import { EventEspressoDomData } from '../../../application/DomDataTypes';
 import { CurrentUserProps, GeneralSettings } from '../../../application/valueObjects/config/types';
 import { DatetimeEdge, TicketEdge, PriceEdge, PriceTypeEdge, Event } from '../services/apollo/types';
 
@@ -9,7 +9,7 @@ export interface EditorData {
 	generalSettings?: GeneralSettings;
 }
 
-export interface EventData {
+export interface EventData extends Event {
 	dbId: number;
 	datetimes?: DatetimeEdge;
 	tickets?: TicketEdge;
@@ -18,14 +18,8 @@ export interface EventData {
 	relations?: RelationalData;
 }
 
-// All of the above defs will be removed
 export interface EventEditorData {
-	event: Event;
-	datetimes?: DatetimeEdge;
-	tickets?: TicketEdge;
-	prices?: PriceEdge;
-	priceTypes?: PriceTypeEdge;
-	relations?: RelationalData;
+	event: EventData;
 }
 
 export interface EventEditorDomData extends EventEspressoDomData {
